+++ docs/changelog.rst
@@ -17,6 +17,8 @@ Versions follow `CalVer`_ with a strict backwards compatibility policy. The thir
 - Added :func:`~forge.fsignature` which is a convenience for :func:`~forge.FSignature.from_callable` to reflect the functionality of :func:`inspect.signature` (itself a convenience for :func:`inspect.Signature.from_callable`)
 - Added slice notation support for :class:`~forge.FSignature` objects.
 - Exposed :class:`~forge.CallArguments`
+- Removed ``get_return_type`` and ``set_return_type`` as :attr:`~forge.FSignature.return_annotation` is now a first-class attribute
+- Renamed stringify_callable -> repr_callable
 
 .. _changelog_2018-5-0: